Refactor device.rs into smaller files
